App/deployment

+++ backend/src/main/java/com/ku/quizzical/app/controller/user/UserController.java
@@ -10,59 +10,76 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
+import com.ku.quizzical.app.helper.AuthorizationValidationHelper;
+import com.ku.quizzical.app.helper.BackendValidationHelper;
 
 @CrossOrigin
 @RestController
-@RequestMapping("/users")
 public final class UserController {
     // Instance Fields
-    private UserServiceImpl service;
+    private UserDatabaseService service;
+    private UserRepository repository;
 
     // Constructor Method
-    public UserController(UserServiceImpl service) {
+    public UserController(UserDatabaseService service, UserRepository repository) {
         super();
         this.service = service;
+        this.repository = repository;
     }
 
     // CREATE Method
     // Saves a User
-    @PostMapping()
-    public ResponseEntity<UserDto> saveUser(@RequestBody UserRegistrationRequest question) {
-        return new ResponseEntity<UserDto>(this.service.saveUser(question), HttpStatus.OK);
+    @PostMapping("/users")
+    public ResponseEntity<UserDto> saveUser(@RequestBody UserRegistrationRequest user) {
+        return new ResponseEntity<UserDto>(this.service.saveUser(user), HttpStatus.OK);
     }
 
     // READ Method
     // Gets all Users
-    @GetMapping()
+    @GetMapping("/users")
     public List<UserDto> getAllUsers() {
         return this.service.getAllUsers();
     }
 
     // READ Method
     // Gets a User by its id
-    @GetMapping("{id}")
+    @GetMapping("/users/{id}")
     public ResponseEntity<UserDto> getUserById(@PathVariable String id) {
-        return new ResponseEntity<UserDto>(this.service.getUserById(id), HttpStatus.OK);
+        return new ResponseEntity<UserDto>(this.service.getUser(id), HttpStatus.OK);
+    }
+
+    // READ Method
+    // Gets a User by its username
+    @GetMapping("/users/username/{username}")
+    public ResponseEntity<UserDto> getUserByUsername(@PathVariable String username) {
+        return new ResponseEntity<UserDto>(this.service.getUserByUsername(username), HttpStatus.OK);
     }
 
     // UPDATE Method
     // Updates a User
-    @PatchMapping("{id}")
-    public ResponseEntity<UserDto> updateUser(@PathVariable String id, @RequestBody UserDto user) {
-        return new ResponseEntity<UserDto>(this.service.updateUser(user, id), HttpStatus.OK);
+    @PatchMapping("/users/{id}")
+    public ResponseEntity<UserDto> updateUser(@PathVariable String id,
+            @RequestHeader("Authorization") String authorizationHeader,
+            @RequestBody UserUpdateRequest user) {
+        User matchingUser = BackendValidationHelper.validateExistingResourceWithFallthrough("User",
+                id, this.repository::findById);
+        AuthorizationValidationHelper.validateAuthorization(authorizationHeader,
+                matchingUser.getId());
+        return new ResponseEntity<UserDto>(this.service.updateUser(id, user), HttpStatus.OK);
     }
 
     // DELETE Method
     // Deletes a User
-    @DeleteMapping("{id}")
-    public String deleteUser(@PathVariable String id) {
-        try {
-            this.service.deleteUser(id);
-        } catch (Exception e) {
-            // Do Nothing
-        }
+    @DeleteMapping("/users/{id}")
+    public String deleteUser(@PathVariable String id,
+            @RequestHeader("Authorization") String authorizationHeader) {
+        User matchingUser = BackendValidationHelper.validateExistingResourceWithFallthrough("User",
+                id, this.repository::findById);
+        AuthorizationValidationHelper.validateAuthorization(authorizationHeader,
+                matchingUser.getId());
+        this.service.deleteUser(id);
         return "\"The User with id \\\"" + id + "\\\" has been deleted.\"";
     }
 }
